requirements.txt
fastapi
uvicorn[standard]
openai
langchain
pinecone-client
unstructured
pdfminer.six
tiktoken